Introducing Use-Auto transform for cpp11-migrate
The new Use-Auto transform replaces the type specifier for variable
declarations with the special C++11 'auto' type specifier. For now, the
replacement is done only for variables that are iterators of any of the
std containers and only if the type used is one of those explicitly
allowed by the standard (i.e. not an implementation-specific type).

Reviewers: gribozavr, silvas, klimek


git-svn-id: http://llvm.org/svn/llvm-project/clang-tools-extra/trunk@176266 91177308-0d34-0410-b5e6-96231b3b80d8