Test case for PR-15414 - nullptr_t functions
The use-null-ptr transform will transform calls to functions that return a
nullptr_t. Even if the function were to only return a null pointer and do
nothing else, this replacement would still be undesired as the behavior and
signature of the function could change in the future.

This adds an XFAILed test case to demonstrate the issue.

Reviewed by: Edwin Vane, Tareq Siraj



git-svn-id: http://llvm.org/svn/llvm-project/clang-tools-extra/trunk@176553 91177308-0d34-0410-b5e6-96231b3b80d8