cpp11-migrate: Fix build break on Windows introduced in r187041.
r187041 changed the way the transform are created and the order the transformed
are applied "may" have changed as well (no specific constraints exist on the
order). This produced a test failure on the Windows buildbot.

Now the test use -std=c++11, so 'nullptr' is defined and the test is not anymore
dependent on the order the transforms are applied.

git-svn-id: http://llvm.org/svn/llvm-project/clang-tools-extra/trunk@187048 91177308-0d34-0410-b5e6-96231b3b80d8